% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istanbul.R
\name{add_istanbul_part}
\alias{add_istanbul_part}
\title{Add Istanbul continent part to city}
\usage{
add_istanbul_part(frame)
}
\arguments{
\item{frame}{A frame with \code{city} and \code{town} variables.}
}
\description{
Given a frame with \code{city} and \code{town} variables, if there are
towns for Istanbul, it add the continent information to the \code{city}
variable of that town.
}

