% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{cache}
\alias{cache}
\title{Taken from https://github.com/hadley/dplyr
/blob/f45629630b0b55452e752a3e4c9948fbb6226ca1/R/data.r}
\format{An object of class \code{environment} of length 0.}
\usage{
cache
}
\description{
Taken from https://github.com/hadley/dplyr
/blob/f45629630b0b55452e752a3e4c9948fbb6226ca1/R/data.r
}
\keyword{datasets}

