% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_computation}
\alias{cache_computation}
\title{Caching computation in a local environment}
\usage{
cache_computation(name, computation)
}
\description{
Computation is assigned a key, and everytime it's called, it is retrieed
from a cache in local environment. This is different than memoise because
connections can also be cached since they are put inside and environemnt.
}

