% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{db_location}
\alias{db_location}
\title{Check that database location is writable}
\usage{
db_location(filename, path = NULL, pkg_name = NULL)
}
\arguments{
\item{path}{if there is a specific path for the db, this parameter is used}

\item{pkg_name}{if there is no path, root of the pkg_name package is used}
}
\description{
Check that database location is writable
}

