% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{discard_outliers}
\alias{discard_outliers}
\title{Discard outliers from a frame given value variable}
\usage{
discard_outliers(frame, value_var, percentile = 0.025)
}
\arguments{
\item{frame}{\code{data.frame} variant dataset.}

\item{value_var}{name of the value variable that will be
inspected for outliers}

\item{value_var}{name of the value variable}

\item{percentage}{what percentile will be considered outlier. This will be
used for upper and lower percentile.}
}
\description{
Some entries in datasets are outliers. This function filters them out.
Given frame can be grouped.
}

