% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_merged_cells.R
\name{fill_merged_cells}
\alias{fill_merged_cells}
\title{Fill merged cell with previous value}
\usage{
fill_merged_cells(vec)
}
\arguments{
\item{vec}{A vector that is known to have merged cells}
}
\value{
\code{vec} with \code{NA}'s filled with previous cell value
}
\description{
Sometimes in an excel sheet or a csv, some cells of a column are merged.
Reader can understand what is meant on a proper viewer but when the data
is read into R, only one of the merged cell has data and others are \code{NA}
}
\details{
This function keeps the record of last cell with a value and if the next cell
is \code{NA}, puts the last seen value in previous cells.

\strong{CAUTION!} If the cell is actually supposed to be \code{NA}, it still
fills is with the last seen value.
}
\examples{
x <- c("A", NA, NA, "B", NA)
fill_merged_cell(x)

}

