% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regex_pattern.R
\name{get_regex_pattern}
\alias{get_regex_pattern}
\title{Extract regex pattern from character vectors}
\usage{
get_regex_pattern(char_vector, pattern)
}
\arguments{
\item{char_vector}{Character vector that the regex will be extracted from}

\item{pattern}{Regex pattern to extract}
}
\value{
Extracted regex patterns as a character vector
}
\description{
Given a set of character vector, extracts the given regex pattern out
of every single element of that vector.
}
\examples{
# Getting digits out of name-digit combinations
x <- c("John-1", "Jess-2", "Peter-4")
get_regex_pattern(x, "[A-Za-z]+-([0-9]+)")

}

