% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_frame_utf8.R
\name{make_frame_utf8}
\alias{make_frame_utf8}
\title{Changing the encoding of frame columns to utf8}
\usage{
make_frame_utf8(frame)
}
\arguments{
\item{frame}{an object inheriting data.frame}
}
\value{
inputted frame with character vectors encoded as utf8
}
\description{
Encoding differences can make joins not work properly. This function
get a variance of data.frame as input and converts the encoding of every
character column in that frame to utf8.
}
\examples{
# Creating strings encoded latin1
x <- c("fa\\xe7ile", "\\xe7erez")
Encoding(x) <- "latin1"

# Putting them in a data.frame
df <- data.frame(x = x, y = c(1, 2))
make_frame_utf8(df)

}

