% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{message_time}
\alias{message_time}
\title{Print message with the current system time}
\usage{
message_time(...)
}
\description{
This is a wrapper around \code{\link{message}}. The only difference is
it passes the current system time as the first argument
}

