% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_na.R
\name{null_to_na}
\alias{null_to_na}
\title{Changing occurances of NULL to NA}
\usage{
null_to_na(vector)
}
\description{
This is usefull if you want to simplify a list into a
built in vector and use it in a data_frame.
}

