% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_update_fn.R
\name{status_update_fn_generator}
\alias{status_update_fn_generator}
\title{A utility function for progress}
\usage{
status_update_fn_generator(identifier, current_count, end_count)
}
\arguments{
\item{identifier}{what the process is about}

\item{end_count}{how many things will have to be done eventually}

\item{start_count}{how many of the things trying to be done is already done}
}
\value{
a function that takes the current count as an argument. When it is called,
     it prints a status message
}
\description{
This function makes a linear model prints a status
message about the process that is tring to be completed.
}

