% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strptime_tr.R
\name{strptime_tr}
\alias{strptime_tr}
\title{strptime in TR locale}
\usage{
strptime_tr(x, format, tz = "")
}
\value{
\link[base]{POSIXlt}
}
\description{
This function is a mirror for strptime function. The only difference
is that it sets the locale to tr_TR before calling it. After completion,
it sets the locale back to whatever it was before.
}

