% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_currency_rates.R
\docType{data}
\name{tl_currency_rates}
\alias{tl_currency_rates}
\title{Major currency rates compared to TRY}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.}
\usage{
tl_currency_rates
}
\description{
Recorded as of 05/09/2016
}
\keyword{datasets}

