% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istanbul.R
\docType{data}
\name{town_to_istanbul_part}
\alias{town_to_istanbul_part}
\title{Dataset that matches Istanbul towns with the continent part}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 39 rows and 2 columns.}
\usage{
town_to_istanbul_part
}
\description{
Dataset that matches Istanbul towns with the continent part
}
\keyword{datasets}

