% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_char_to_ascii.R
\name{tr_char_to_ascii}
\alias{tr_char_to_ascii}
\title{Changing turkish characters to ascii counterparts}
\usage{
tr_char_to_ascii(char_vector)
}
\value{
character vector without Turkish characters
}
\description{
Changing turkish characters to ascii counterparts
}
\examples{
tr_char_to_ascii(c("Şekil", "Kağnı", "Otobüs"))

}

